requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
python-dotenv
beautifulsoup4
faiss-cpu
streamlit
PyPDF2
pypdf
langchain-ollama
aiohttp